+++ src/color/color.h
@@ -8,7 +8,7 @@
 
 typedef t_tuple t_color;
 
-t_color		ft_color(float r, float g, float b);
+t_color		ft_color(double r, double g, double b);
 t_color		ft_hexcolor(uint32_t color);
 uint32_t	to_raw_color(t_color c);
 bool 		color_equal(t_color c1, t_color c2);
