+++ src/raytracer/raytracer.h
@@ -26,7 +26,7 @@ struct s_ray
 typedef struct s_inter t_inter;
 struct s_inter
 {
-	float		t;
+	double		t;
 	t_object	object;
 };
 
@@ -41,7 +41,7 @@ struct s_xs
 typedef struct s_computations t_computations;
 struct s_computations
 {
-	float		t;
+	double		t;
 	t_object	object;
 	t_tuple		point;
 	t_tuple		eyev;
@@ -56,7 +56,7 @@ t_light		*get_light_at(t_list *lst, int index);
 
 
 // globals
-t_inter		ft_inter(float t, t_object object);
+t_inter		ft_inter(double t, t_object object);
 t_xs		ft_inters(unsigned int count, ...);
 t_xs		intersect(t_object object, t_ray ray);
 t_tuple		normal_at(t_object object, t_tuple p);
@@ -66,7 +66,7 @@ void		destroy_inters(t_xs *xs);
 
 // ray
 t_ray			ft_ray(t_tuple origin, t_tuple direction);
-t_tuple			ft_position(t_ray r, float t);
+t_tuple			ft_position(t_ray r, double t);
 t_ray			ray_transform(t_ray ray, t_matrix transformation_matrix);
 
 // sphere
@@ -82,10 +82,10 @@ bool			is_shadowed(t_scene s, t_light light, t_tuple point);
 // material
 t_material		ft_material();
 bool			material_equal(t_material m1, t_material m2);
-void			setMatAmbient(t_material *mat, float ambient);
-void			setMatDiffuse(t_material *mat, float diffuse);
-void			setMatSpecular(t_material *mat, float specular);
-void			setMatShininess(t_material *mat, float shininess);
+void			setMatAmbient(t_material *mat, double ambient);
+void			setMatDiffuse(t_material *mat, double diffuse);
+void			setMatSpecular(t_material *mat, double specular);
+void			setMatShininess(t_material *mat, double shininess);
 void			print_material(t_material m);
 
 // world
@@ -100,7 +100,7 @@ t_color			shade_hit(t_scene w, t_computations comps);
 t_computations	prepare_computations(t_inter intersection, t_ray r);
 
 // camera
-t_ray			ray_for_pixel(t_camera c, float x, float y);
+t_ray			ray_for_pixel(t_camera c, double x, double y);
 t_canvas		render(t_camera c, t_scene s);
 void			print_camera_info(t_camera c);
