+++ src/tuple/tuple.h
@@ -8,36 +8,36 @@ union u_tuple
 {
 	struct
 	{
-		float x;
-		float y;
-		float z;
-		float w;
+		double x;
+		double y;
+		double z;
+		double w;
 	};
 	struct
 	{
-		float r;
-		float g;
-		float b;
-		float a;
+		double r;
+		double g;
+		double b;
+		double a;
 	};
 };
 
 
-t_tuple	point(float x, float y, float z);
-t_tuple	vector(float x, float y, float z);
-t_tuple tuple(float x, float y, float z, float w);
+t_tuple	point(double x, double y, double z);
+t_tuple	vector(double x, double y, double z);
+t_tuple tuple(double x, double y, double z, double w);
 
 bool	is_equal(t_tuple t1, t_tuple t2);
 
 t_tuple	add_tuple(t_tuple pv, t_tuple v);
 t_tuple	subst_tuple(t_tuple p, t_tuple pv);
 t_tuple	negate_tuple(t_tuple v);
-t_tuple	multiply_tuple(t_tuple v, float scalar);
-t_tuple	divide_tuple(t_tuple v, float scalar);
+t_tuple	multiply_tuple(t_tuple v, double scalar);
+t_tuple	divide_tuple(t_tuple v, double scalar);
 t_tuple	tuple_product(t_tuple t1, t_tuple t2);
 
-float	dot(t_tuple v1, t_tuple v2);
-float	magnitude(t_tuple v);
+double	dot(t_tuple v1, t_tuple v2);
+double	magnitude(t_tuple v);
 t_tuple	normalize(t_tuple v);
 t_tuple	cross(t_tuple a, t_tuple b);
