+++ src/utils/utils.h
@@ -20,6 +20,6 @@ void	*ft_free(void *p);
 void	*ft_arr_free(char **arr);
 void	arr_print(char **arr);
 bool	is_int(const char *s);
-float	max(float a, float b);
+double	max(double a, double b);
 
 #endif
